\name{banana}
\docType{data}
\alias{banana}
\title{Biompedance measurements on bananas.}
\description{
Bioimpedance measurements and weight of a production batch of 50 bananas observed for 11 days. Bioimpedance has been measured at 3 different electrical frequencies of a portable electrical impedance spectroscopy device, called Fruit Meter. The experiment has been conducted by the Sensing Technologies Laboratory \url{https://sensingtechnologies.groups.unibz.it/} of the Faculty of Engineering (ex Fac. Science and Technology) at the Free University of Bozen-Bolzano (Unibz, Italy) in the context of an interdisciplinary research project with the Faculty of Economics and Management (Unibz, Italy).
}
\format{
A matrix containing three-way observations.
}
\references{
Ibba, P., Falco,  A., Abera, B., Cantarella, G., Petti, L., and Lugli, P. (2019). Bio-impedance and circuit parameters: An analysis for tracking fruit ripening. Postharvest Biology and Technology, 159. DOI: 10.1016/j.postharvbio.2019.110978. 

Ibba P., et al. (2021). Design and Validation of a Portable AD5933-Based Impedance Analyzer for Smart Agriculture, in IEEE Access, 9, pp. 63656-63675. DOI: 10.1109/ACCESS.2021.3074269.

Ibba, P. (2021). Fruit quality evaluation using electrical impedance spectroscopy. PhD thesis, 196, Supervisor: Lugli P., Free University of Bozen-Bolzano.
}
\keyword{datasets}
